bug fix activation token
