+++ FRONTEND/src/app/components/auth/auth-activation/auth-activation.component.html
@@ -1,0 +1,19 @@
+<div class="activation-page text-center p-5">
+  <ng-container *ngIf="status === 'pending'">
+    Activation en cours…
+  </ng-container>
+
+  <ng-container *ngIf="status === 'success'">
+    <h3 class="text-success">✅ {{ message }}</h3>
+    <button class="btn btn-primary mt-3" routerLink="/connexion">
+      Se connecter
+    </button>
+  </ng-container>
+
+  <ng-container *ngIf="status === 'error'">
+    <h3 class="text-danger">❌ {{ message }}</h3>
+    <button class="btn btn-secondary mt-3" routerLink="/inscription">
+      Réessayer l’inscription
+    </button>
+  </ng-container>
+</div>
